requirements.txt
albumentations==1.3.0
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attr==0.3.1
attrs==22.2.0
azure-core==1.26.2
azure-storage-blob==12.14.1
bleach==5.0.1
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.5
cryptography==39.0.0
cycler==0.11.0
dataclasses==0.6
defusedxml==0.7.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
elasticdeform==0.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
expiringdict==1.2.2
fonttools==4.37.1
google-api-core==2.10.0
google-auth==2.11.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.3
grpcio==1.51.1
grpcio-status==1.48.2
htmlmin==0.1.12
idna==3.3
imageio==2.21.3
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.2.0
kiwisolver==1.4.4
launchdarkly-server-sdk==7.5.0
lazy_loader==0.1rc2
littleutils==0.2.2
lockfile==0.12.2
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.5.3
medutils==0.1.19
monai==1.0.1
msrest==0.7.1
munch==2.5.0
networkx==2.8.6
nibabel==5.1.0
nltk==3.6.7
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
ordered-set==4.0.2
outdated==0.2.2
pandas==1.5.2
pandas-flavor==0.3.0
parameterized==0.8.1
patsy==0.5.3
Pillow==9.3.0
pingouin==0.5.3
pluggy==1.0.0
proto-plus==1.22.2
protobuf==3.20.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydicom==2.3.0
PyJWT==2.4.0
pyRFC3339==1.1
pyrsistent==0.19.3
#PySimpleGUI==4.60.4
pytest==7.2.2
pytest-cov==4.0.0
python-json-logger==2.0.4
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
redis==4.4.2
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rq==1.10.1
rsa==4.9
rt-utils==1.2.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.2
seg-metrics==1.1.3
segmentation-mask-overlay==0.3.4
semver==2.13.0
sentry-sdk==1.14.0
SimpleITK==2.2.1
simplejson==3.18.1
sqlparse==0.4.3
statsmodels==0.13.5
tabulate==0.9.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tomli==2.0.1
torch==1.12.1
torch-summary==1.4.5
torchvision==0.13.1
tqdm==4.64.1
twilio==7.14.0
typing_extensions==4.3.0
ua-parser==0.16.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
webencodings==0.5.1
xarray==2023.1.0
xmljson==0.2.0
